refactor:  access (#362)
* feat: add role model, handle data migration for existing role field

* feat: clean up legacy role logic from models

* feat: clean up data migration logic

* chore: model update migrations

* feat: infer default role policy from default_roles

* feat: add role type

* feat: add roles query, misc other updates

* refactor: simplify policy structure

* feat: misc updates to mutation resolvers for new role schema

* feat: update frontend queries, mutations, schema and types

* feat: add frontend utils to gatekeep actions based on permission policy

* feat: update frontend components to use rbac system

* fix: add id to roles

* fix: mutation arg

* fix: use role name for legacy util

* feat: add integration credentials permission category to default roles

* fix: update user role update logic

* fix: replace naive permission checks with granular permissions, fix role label props

* fix: misc fixes for granular rbac

* feat: formalize global access policy

* fix: only allow users with access to all envs to enable sse

* feat: access page, custom roles (#367)

* feat: access page, custom roles

* fix: save button disabled state

* chore: add common util to handle permission policy updates

* fix: misc styling updates to access layout

* fix: sidebar link

* fix: infer colors for custom roles

* fix: conditional rendering, manage role disabled states

* fix: misc cleanup

* fix: sidebar links order

* feat: add accordions for permission tables, misc ui updates to role dialogs

* feat: enforce read permission for apps

* fix: update app members manage dialog to handle global access setting

* feat: add color field to role model

* feat: add meta field to default roles

* fix: get default role description from meta field

* feat: add color to role mutations

* feat: add color to client-side graphql mutations

* feat: misc updates to create and manage role dialogs

* chore: misc fixes to client side queries

* fix: correctly check user global access config

* fix: restyle user roles list

* fix: htmlfor attr in input component

* feat: add common textarea component

* feat: move random hexcolor util to common utils

* feat: size variants to role label, misc updates

* feat: add access level templates, misc other fixes

* feat: add color picker with curated colors

* fix: disable color picker when edit not allowed

* feat: move colors to copy utils, add util to get random color

* feat: show selected color

* feat: use curated random color for new role color state

* feat: changed name to Access Control

* fix: command palette navigation

* feat: updated role info description

* fix: remove label from color picker component

* fix: sanitize role name to include non-space chars

* fix: color picker disabled prop

* fix: show action buttons on hover, add title to locked role icon

---------


* fix: closing tag

* feat: add lockbox to app permissions

* feat: explicitly store and read app encryption mode (#373)

* feat: add explicit model field for app encryption mode

* fix: permission check for enable sse dialog in app settings

* fix: remove legacy empty state

* fix: handle env index edge case

* chore: misc cleanup

* feat: enforce granular permissions (#369)

* feat: enforce org, billing, app permissions

* feat: enforce permissions for users and roles

* feat: enforce permissions for integrations and credentials

* feat: enforce permissions for secrets and environments

* feat: conditionally render encryption mode based on permissions

* feat: poll organisation memberships and permissions

* fix: loading states, permissions for new secrets, syncs

* feat: enforce permissions for app members

* fix: check permissions when resolving secret event user

* feat: enforce permissions for logs

* feat: enforce permissions for service tokens

* feat: enforce permissions for lockbox

* fix: args for app permission checks

* fix: check env and secret create permissions for new app flow

* fix: enforce and handle permissions for org home screen

* fix: members card link

* fix: check for app permission when fetching org syncs

* fix: return empty list instead of error if apps:read permission isn't available

* fix: disabled prop for global access toggle

* feat: prevent users without global access updating a user's role to one with ga

* fix: remove ga from custom roles

* feat: add descriptor for ga to role label descriptions

* fix: add no-permission empty state for apps

* fix: org members link

* fix: misc fixes for secrets permissions

* fix: misc fixes to REST api error responses

* fix: enforce permissions for folder deletes

* fix: enforce env read permissions when adding app members

* feat: rename service creds to integration creds

* fix: horizontal padding on settings page

* fix: role selector dropdown width

* fix: enforce envs-read permission when creating integration

* fix: gatekeeping copy in manage env dialog

* feat: misc updates

* fix: orm query in permission util

---------

+++ backend/backend/graphene/mutations/organisation.py
@@ -1,6 +1,10 @@
 from api.emails import send_invite_email, send_user_joined_email, send_welcome_email
-from api.utils.permissions import user_is_admin, user_is_org_member
-
+from api.utils.access.permissions import (
+    user_has_permission,
+    user_is_admin,
+    user_is_org_member,
+)
+from api.utils.access.roles import default_roles
 import graphene
 from graphql import GraphQLError
 from api.models import (
@@ -9,6 +13,7 @@ from api.models import (
     CustomUser,
     OrganisationMember,
     OrganisationMemberInvite,
+    Role,
 )
 from backend.graphene.types import (
     OrganisationMemberInviteType,
@@ -39,10 +44,20 @@ class CreateOrganisationMutation(graphene.Mutation):
 
         user = CustomUser.objects.get(userId=info.context.user.userId)
         org = Organisation.objects.create(id=id, name=name, identity_key=identity_key)
+
+        for role_name, _ in default_roles.items():
+            Role.objects.create(
+                name=role_name,
+                organisation=org,
+                is_default=True,
+            )
+
+        owner_role = Role.objects.get(organisation=org, name__iexact="owner")
+
         owner = OrganisationMember.objects.create(
             user=user,
             organisation=org,
-            role=OrganisationMember.OWNER,
+            role=owner_role,
             identity_key=identity_key,
             wrapped_keyring=wrapped_keyring,
             wrapped_recovery=wrapped_recovery,
@@ -92,14 +107,16 @@ class InviteOrganisationMemberMutation(graphene.Mutation):
         org_id = graphene.ID(required=True)
         email = graphene.String(required=True)
         apps = graphene.List(graphene.String)
-        role = graphene.String()
 
     invite = graphene.Field(OrganisationMemberInviteType)
 
     @classmethod
-    def mutate(cls, root, info, org_id, email, apps, role):
+    def mutate(cls, root, info, org_id, email, apps):
 
         org = Organisation.objects.get(id=org_id)
+
+        if not user_has_permission(info.context.user, "create", "Members", org):
+            raise GraphQLError("You dont have permission to invite members")
 
         if user_is_org_member(info.context.user, org_id):
             user_already_exists = OrganisationMember.objects.filter(
@@ -129,7 +146,6 @@ class InviteOrganisationMemberMutation(graphene.Mutation):
             invite = OrganisationMemberInvite.objects.create(
                 organisation=org,
                 invited_by=invited_by,
-                role=role.lower(),
                 invitee_email=email,
                 expires_at=expiry,
             )
@@ -155,6 +171,11 @@ class DeleteInviteMutation(graphene.Mutation):
     @classmethod
     def mutate(cls, rooot, info, invite_id):
         invite = OrganisationMemberInvite.objects.get(id=invite_id)
+
+        if not user_has_permission(
+            info.context.user, "delete", "Members", invite.organisation
+        ):
+            raise GraphQLError("You dont have permission to delete invites")
 
         if user_is_org_member(info.context.user, invite.organisation.id):
             invite.delete()
@@ -198,10 +219,12 @@ class CreateOrganisationMemberMutation(graphene.Mutation):
 
             org = Organisation.objects.get(id=org_id)
 
+            dev_role = Role.objects.get(organisation=org, name__iexact="developer")
+
             org_member = OrganisationMember.objects.create(
                 user_id=info.context.user.userId,
                 organisation=org,
-                role=invite.role,
+                role=dev_role,
                 identity_key=identity_key,
                 wrapped_keyring=wrapped_keyring,
                 wrapped_recovery=wrapped_recovery,
@@ -238,6 +261,11 @@ class DeleteOrganisationMemberMutation(graphene.Mutation):
     def mutate(cls, root, info, member_id):
         org_member = OrganisationMember.objects.get(id=member_id, deleted_at=None)
 
+        if not user_has_permission(
+            info.context.user, "delete", "Members", org_member.organisation
+        ):
+            raise GraphQLError("You dont have permission to remove members")
+
         if org_member.user == info.context.user:
             raise GraphQLError("You can't remove yourself from an organisation")
 
@@ -257,21 +285,25 @@ class DeleteOrganisationMemberMutation(graphene.Mutation):
 class UpdateOrganisationMemberRole(graphene.Mutation):
     class Arguments:
         member_id = graphene.ID(required=True)
-        role = graphene.String(required=True)
+        role_id = graphene.ID(required=True)
 
     org_member = graphene.Field(OrganisationMemberType)
 
     @classmethod
-    def mutate(cls, root, info, member_id, role):
+    def mutate(cls, root, info, member_id, role_id):
         org_member = OrganisationMember.objects.get(id=member_id, deleted_at=None)
 
-        if user_is_admin(info.context.user.userId, org_member.organisation.id):
-            if role.lower() == OrganisationMember.OWNER.lower():
-                raise GraphQLError("You cannot set this user as the organisation owner")
+        if not user_has_permission(
+            info.context.user, "update", "Members", org_member.organisation
+        ):
+            raise GraphQLError("You dont have permission to change member roles")
 
-            org_member.role = role.lower()
-            org_member.save()
+        role = Role.objects.get(organisation=org_member.organisation, id=role_id)
 
-            return UpdateOrganisationMemberRole(org_member=org_member)
-        else:
-            raise GraphQLError("You don't have permission to perform this action")
+        if role.name.lower() == "owner":
+            raise GraphQLError("You cannot set this user as the organisation owner")
+
+        org_member.role = role
+        org_member.save()
+
+        return UpdateOrganisationMemberRole(org_member=org_member)
